refactor cors logic into middleware folder
