+++ backend/config/config.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/gin-contrib/cors"
 	redigo "github.com/gomodule/redigo/redis"
 	"github.com/joho/godotenv"
 )
@@ -21,7 +20,6 @@ type Config struct {
 	Database      DatabaseConfig
 	Redis         RedisConfig
 	OIDC          OIDCConfig
-	CORS          CORSConfig
 	JWT           JWTConfig
 	SessionSecret string
 }
@@ -132,17 +130,6 @@ type OIDCConfig struct {
 	PostLogoutRedirectURL string
 }
 
-// CORSConfig holds CORS-related configuration.
-type CORSConfig struct {
-	AllowOrigins     []string
-	AllowAllOrigins  bool
-	AllowMethods     []string
-	AllowHeaders     []string
-	ExposeHeaders    []string
-	AllowCredentials bool
-	MaxAge           int
-}
-
 // Load reads configuration from environment variables.
 // .env file is optional (useful for local development, not needed in Docker)
 func Load() (*Config, error) {
@@ -178,9 +165,6 @@ func Load() (*Config, error) {
 	}
 	endSessionURL := strings.TrimSpace(os.Getenv("OIDC_END_SESSION_URL"))
 	postLogoutRedirect := strings.TrimSpace(os.Getenv("OIDC_POST_LOGOUT_REDIRECT_URL"))
-
-	// CORS configuration
-	corsConfig := loadCORSConfig()
 
 	// Database configuration
 	dbConfig := DatabaseConfig{
@@ -231,98 +215,12 @@ func Load() (*Config, error) {
 			EndSessionURL:         endSessionURL,
 			PostLogoutRedirectURL: postLogoutRedirect,
 		},
-		CORS: corsConfig,
 		JWT: JWTConfig{
 			Secret:   jwtSecret,
 			Duration: jwtDuration,
 		},
 		SessionSecret: sessionSecret,
 	}, nil
-}
-
-// loadCORSConfig loads CORS configuration from environment variables.
-func loadCORSConfig() CORSConfig {
-	// Allow all origins by default (can be restricted via CORS_ALLOW_ORIGINS)
-	allowAllOrigins := getEnv("CORS_ALLOW_ALL_ORIGINS", "true") == "true"
-
-	var allowOrigins []string
-	if !allowAllOrigins {
-		originsStr := getEnv("CORS_ALLOW_ORIGINS", "")
-		if originsStr != "" {
-			allowOrigins = strings.Split(originsStr, ",")
-			// Trim whitespace
-			for i := range allowOrigins {
-				allowOrigins[i] = strings.TrimSpace(allowOrigins[i])
-			}
-		}
-	}
-
-	// Allow methods
-	methodsStr := getEnv("CORS_ALLOW_METHODS", "GET,POST,PUT,PATCH,DELETE,HEAD,OPTIONS")
-	allowMethods := strings.Split(methodsStr, ",")
-	for i := range allowMethods {
-		allowMethods[i] = strings.TrimSpace(allowMethods[i])
-	}
-
-	// Allow headers
-	headersStr := getEnv("CORS_ALLOW_HEADERS", "Origin,Content-Type,Accept,Authorization")
-	allowHeaders := strings.Split(headersStr, ",")
-	for i := range allowHeaders {
-		allowHeaders[i] = strings.TrimSpace(allowHeaders[i])
-	}
-
-	// Expose headers
-	exposeStr := getEnv("CORS_EXPOSE_HEADERS", "Content-Length")
-	exposeHeaders := strings.Split(exposeStr, ",")
-	for i := range exposeHeaders {
-		exposeHeaders[i] = strings.TrimSpace(exposeHeaders[i])
-	}
-
-	// Allow credentials
-	allowCredentials := getEnv("CORS_ALLOW_CREDENTIALS", "true") == "true"
-
-	// Max age (in seconds, default 12 hours)
-	maxAgeStr := getEnv("CORS_MAX_AGE", "43200")
-	maxAge, _ := strconv.Atoi(maxAgeStr)
-	if maxAge <= 0 {
-		maxAge = 43200 // 12 hours default
-	}
-
-	return CORSConfig{
-		AllowOrigins:     allowOrigins,
-		AllowAllOrigins:  allowAllOrigins,
-		AllowMethods:     allowMethods,
-		AllowHeaders:     allowHeaders,
-		ExposeHeaders:    exposeHeaders,
-		AllowCredentials: allowCredentials,
-		MaxAge:           maxAge,
-	}
-}
-
-// GetCORSConfig returns the gin-contrib/cors Config based on the CORSConfig.
-func (c *CORSConfig) GetCORSConfig() cors.Config {
-	config := cors.Config{
-		AllowMethods:     c.AllowMethods,
-		AllowHeaders:     c.AllowHeaders,
-		ExposeHeaders:    c.ExposeHeaders,
-		AllowCredentials: c.AllowCredentials,
-		MaxAge:           time.Duration(c.MaxAge) * time.Second,
-	}
-
-	if c.AllowAllOrigins {
-		config.AllowOriginFunc = func(origin string) bool {
-			return true
-		}
-	} else if len(c.AllowOrigins) > 0 {
-		config.AllowOrigins = c.AllowOrigins
-	} else {
-		// Default: allow all if nothing specified
-		config.AllowOriginFunc = func(origin string) bool {
-			return true
-		}
-	}
-
-	return config
 }
 
 // getEnv returns the value of an environment variable or a default value.
